+++ internal/builtin/register_test.go
@@ -27,6 +27,7 @@ func TestRegister(t *testing.T) {
 	req := registry.Enable(runtime)
 	modules := []string{
 		"osm:argv",
+		"osm:crypto",
 		"osm:fetch",
 		"osm:flag",
 		"osm:nextIntegerId",